requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
audioop-lts==0.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
dataclasses-json==0.6.7
elevenlabs==2.27.0
faiss-cpu==1.13.2
filelock==3.20.1
frozenlist==1.8.0
fsspec==2025.12.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.2.0
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.5
langchain-huggingface==1.2.0
langchain-text-splitters==1.1.0
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.1
langsmith==0.5.1
llvmlite==0.46.0
MarkupSafe==3.0.3
marshmallow==3.26.2
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
networkx==3.6.1
numba==0.63.1
numpy==2.3.5
openai-whisper==20250625
orjson==3.11.5
ormsgpack==1.12.1
packaging==25.0
propcache==0.4.1
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
python-dotenv==1.2.1
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.16.3
sentence-transformers==5.2.0
setuptools==80.9.0
sounddevice==0.5.3
soundfile==0.13.1
SpeechRecognition==3.14.4
SQLAlchemy==2.0.45
standard-aifc==3.13.0
standard-chunk==3.13.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
torch==2.9.1
tqdm==4.67.1
transformers==4.57.3
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
uuid_utils==0.12.0
websockets==15.0.1
wheel==0.45.1
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
